+++ deltafi-core/src/main/java/org/deltafi/core/plugin/deployer/K8sDeployerService.java
@@ -19,6 +19,8 @@ package org.deltafi.core.plugin.deployer;
 
 import io.fabric8.kubernetes.api.model.*;
 import io.fabric8.kubernetes.api.model.apps.Deployment;
+import io.fabric8.kubernetes.api.model.apps.DeploymentCondition;
+import io.fabric8.kubernetes.api.model.apps.DeploymentStatus;
 import io.fabric8.kubernetes.client.KubernetesClient;
 import io.fabric8.kubernetes.client.KubernetesClientTimeoutException;
 import io.fabric8.kubernetes.client.utils.Serialization;
@@ -29,6 +31,7 @@ import org.deltafi.core.plugin.deployer.customization.PluginCustomization;
 import org.deltafi.core.plugin.deployer.customization.PluginCustomizationService;
 import org.deltafi.core.plugin.deployer.image.PluginImageRepository;
 import org.deltafi.core.plugin.deployer.image.PluginImageRepositoryService;
+import org.deltafi.core.services.DeltaFiPropertiesService;
 import org.deltafi.core.services.EventService;
 import org.deltafi.core.snapshot.SystemSnapshotService;
 import org.deltafi.core.types.Result;
@@ -36,32 +39,38 @@ import org.springframework.beans.factory.annotation.Value;
 import org.springframework.core.io.Resource;
 
 import java.io.IOException;
+import java.time.Duration;
 import java.util.ArrayList;
 import java.util.List;
 import java.util.Map;
 import java.util.concurrent.TimeUnit;
-import java.util.stream.Collectors;
 
 @Slf4j
 public class K8sDeployerService extends BaseDeployerService {
-
-    private static final int READY_TIMEOUT = 60;
-    private static final String APP_LABEL_KEY = "app";
-    private static final String PLUGIN_GROUP_LABEL_KEY = "pluginGroup";
-    private static final String CONFIG_MOUNT_NAME = "config";
+    static final String APP_LABEL_KEY = "app";
+    static final String PLUGIN_GROUP_LABEL_KEY = "pluginGroup";
+    static final String CONFIG_MOUNT_NAME = "config";
     private static final String CONFIG_MAP_NAME_TPL = "%s-config";
+    public static final String PROGRESSING = "Progressing";
+    public static final String PROGRESS_DEADLINE_EXCEEDED = "ProgressDeadlineExceeded";
+    private static final String MISSING_POD = "No pod was found for the given plugin coordinates that was not ready";
+
     private final KubernetesClient k8sClient;
+    private final PodService podService;
+    private final DeltaFiPropertiesService deltaFiPropertiesService;
 
     @Value("file:/template/action-deployment.yaml")
     private Resource baseDeployment;
 
-    public K8sDeployerService(PluginImageRepositoryService pluginImageRepositoryService, KubernetesClient k8sClient, PluginCustomizationService pluginCustomizationService, PluginRegistryService pluginRegistryService, SystemSnapshotService systemSnapshotService, EventService eventService) {
+    public K8sDeployerService(DeltaFiPropertiesService deltaFiPropertiesService, PluginImageRepositoryService pluginImageRepositoryService, KubernetesClient k8sClient, PodService podService, PluginCustomizationService pluginCustomizationService, PluginRegistryService pluginRegistryService, SystemSnapshotService systemSnapshotService, EventService eventService) {
         super(pluginImageRepositoryService, pluginRegistryService, pluginCustomizationService, systemSnapshotService, eventService);
         this.k8sClient = k8sClient;
+        this.deltaFiPropertiesService = deltaFiPropertiesService;
+        this.podService = podService;
     }
 
     @Override
-    public Result deploy(PluginCoordinates pluginCoordinates, String imageRepoOverride, String imagePullSecretOverride, String customDeploymentOverride) {
+    public DeployResult deploy(PluginCoordinates pluginCoordinates, String imageRepoOverride, String imagePullSecretOverride, String customDeploymentOverride) {
         PluginImageRepository pluginImageRepository = pluginImageRepositoryService.findByGroupId(pluginCoordinates);
 
         ArrayList<String> info = new ArrayList<>();
@@ -82,11 +91,18 @@ public class K8sDeployerService extends BaseDeployerService {
                     PluginCustomizationService.unmarshalPluginCustomization(customDeploymentOverride) :
                     pluginCustomizationService.getPluginCustomizations(pluginCoordinates);
         } catch (Exception e) {
-            return Result.newBuilder().success(false).info(info).errors(List.of("Could not retrieve plugin customizations: " + e.getMessage())).build();
+            return DeployResult.builder().success(false).info(info).errors(List.of("Could not retrieve plugin customizations: " + e.getMessage())).build();
         }
 
         try {
-            Result result = createOrReplace(createDeployment(pluginCoordinates, pluginImageRepository, pluginCustomization));
+            List<Integer> ports = pluginCustomization != null && pluginCustomization.getPorts() != null ? pluginCustomization.getPorts() : List.of();
+            Deployment deployment = buildDeployment(pluginCoordinates, pluginImageRepository, pluginCustomization, ports);
+            DeployResult result = createOrReplace(deployment, pluginCoordinates);
+            if (result.isSuccess() && !ports.isEmpty()) {
+                // if the deployment was successful and requires a service create it now
+                createOrReplaceService(ports, pluginCoordinates.getArtifactId());
+            }
+
             if(result.getInfo() != null) {
                 result.getInfo().addAll(info);
             }
@@ -95,40 +111,65 @@ public class K8sDeployerService extends BaseDeployerService {
             }
             return result;
         } catch (IOException e) {
-            return Result.newBuilder().success(false).info(info).errors(List.of("Could not create the deployment: " + e.getMessage())).build();
+            return DeployResult.builder().success(false).info(info).errors(List.of("Could not create the deployment: " + e.getMessage())).build();
         }
     }
 
     @Override
-    Result removeDeployment(PluginCoordinates pluginCoordinates) {
-        // TODO: Fix this, it can never be true
-        if (Boolean.FALSE.equals(k8sClient.apps().deployments().withName(pluginCoordinates.getArtifactId()).delete())) {
-            return Result.newBuilder().success(false).errors(List.of("Failed to remove the deployment named " + pluginCoordinates.getArtifactId())).build();
-        }
-
-        return new Result();
+    Result removePluginResources(PluginCoordinates pluginCoordinates) {
+        Result result = new Result();
+        deleteDeployment(pluginCoordinates, result);
+        deleteService(pluginCoordinates, result);
+        return result;
     }
 
     public void setBaseDeployment(Resource resource) {
         baseDeployment = resource;
     }
 
-    private Result createOrReplace(Deployment deployment) {
-        preserveValuesIfUpgrade(deployment, k8sClient.apps().deployments().withName(deployment.getMetadata().getName()).get());
+    private DeployResult createOrReplace(Deployment deployment, PluginCoordinates pluginCoordinates) {
+        Deployment existingDeployment = k8sClient.apps().deployments().withName(deployment.getMetadata().getName()).get();
+
+        boolean isUpgrade = existingDeployment != null;
+
+        if (isUpgrade) {
+            preserveValues(deployment, existingDeployment);
+        }
 
         Deployment installed = k8sClient.resource(deployment).createOrReplace();
 
         try {
-            k8sClient.resource(installed).waitUntilReady(READY_TIMEOUT, TimeUnit.SECONDS);
-        } catch (KubernetesClientTimeoutException timeoutException) {
-            // TODO - should we attempt to rollback/remove the deployment or leave it for further investigation?
-            return Result.newBuilder().success(false).errors(List.of("Deployment " + deployment.getMetadata().getName() + " did not reach a ready state within " + READY_TIMEOUT + " seconds")).build();
+            k8sClient.resource(installed).waitUntilCondition(this::rolloutSuccessful, getTimeoutInMillis(), TimeUnit.MILLISECONDS);
+        } catch (KubernetesClientTimeoutException | IllegalStateException exception) {
+            DeployResult deployResult = new DeployResult();
+            deployResult.setSuccess(false);
+            Pod pod = podService.findNotReadyPluginPod(pluginCoordinates).orElse(null);
+
+            if (pod == null) {
+                deployResult.getErrors().add(MISSING_POD);
+            } else {
+                List<Event> events = podService.podEvents(pod);
+                String logs = podService.podLogs(pod, pluginCoordinates.getArtifactId());
+
+                deployResult.setEvents(events.stream().map(K8sEventUtil::formatEvent).toList());
+                deployResult.setLogs(logs);
+            }
+
+            if (deltaFiPropertiesService.getDeltaFiProperties().getPlugins().isAutoRollback()) {
+                if (isUpgrade) {
+                    k8sClient.apps().deployments().resource(installed).rolling().undo();
+                } else {
+                    k8sClient.apps().deployments().resource(installed).delete();
+                }
+            }
+
+            return deployResult;
         }
 
-        return new Result();
+        return new DeployResult();
     }
 
-    Deployment createDeployment(PluginCoordinates pluginCoordinates, PluginImageRepository pluginImageRepository, PluginCustomization pluginCustomization) throws IOException {
+    Deployment buildDeployment(PluginCoordinates pluginCoordinates, PluginImageRepository pluginImageRepository, PluginCustomization pluginCustomization, List<Integer> ports) throws IOException {
         Deployment deployment = loadBaseDeployment();
 
         String applicationName = pluginCoordinates.getArtifactId();
@@ -145,23 +186,12 @@ public class K8sDeployerService extends BaseDeployerService {
         container.setName(applicationName);
         container.setImage(pluginImageRepository.getImageRepositoryBase() + pluginCoordinates.getArtifactId() + ":" + pluginCoordinates.getVersion());
 
-        if (null != pluginCustomization.getExtraContainers()) {
-            deployment.getSpec().getTemplate().getSpec().getContainers().addAll(pluginCustomization.getExtraContainers());
+        if (!ports.isEmpty()) {
+            container.setPorts(ports.stream().map(this::createContainerPort).toList());
         }
 
-        if (null != pluginCustomization.getPorts()) {
-            List<ContainerPort> containerPorts = pluginCustomization.getPorts().stream().map(this::createContainerPort).map(ContainerPortBuilder::build).collect(Collectors.toList());
-            container.setPorts(containerPorts);
-
-            Service service = new ServiceBuilder().withMetadata(new ObjectMetaBuilder().withName(applicationName).build()).withSpec(
-                    new ServiceSpecBuilder().withPorts(
-                            pluginCustomization.getPorts().stream().map(this::createServicePort).map(b -> b.withTargetPort(new IntOrString(b.getPort()))).map(ServicePortBuilder::build).collect(Collectors.toList())
-                    ).withSelector(
-                            Map.of("app", applicationName)
-                    ).build()
-            ).build();
-            k8sClient.services().resource(service).createOrReplace();
-            log.info("Created service {}", applicationName);
+        if (null != pluginCustomization.getExtraContainers()) {
+            deployment.getSpec().getTemplate().getSpec().getContainers().addAll(pluginCustomization.getExtraContainers());
         }
 
         if (null != pluginImageRepository.getImagePullSecret()) {
@@ -174,18 +204,75 @@ public class K8sDeployerService extends BaseDeployerService {
         return deployment;
     }
 
-    private ContainerPortBuilder createContainerPort(Integer port) {
-        ContainerPortBuilder builder = new ContainerPortBuilder();
-        builder.withContainerPort(port).withName("port-" + port);
-
-        return builder;
+    void createOrReplaceService(List<Integer> ports, String applicationName) {
+        k8sClient.services().resource(buildService(ports, applicationName)).createOrReplace();
     }
 
-    private ServicePortBuilder createServicePort(Integer port) {
-        ServicePortBuilder builder = new ServicePortBuilder();
-        builder.withPort(port);
+    Service buildService(List<Integer> ports, String applicationName) {
+        return new ServiceBuilder()
+                .withNewMetadata()
+                .withName(applicationName)
+                .endMetadata()
+                .withNewSpec()
+                .withSelector(Map.of(APP_LABEL_KEY, applicationName))
+                .withPorts(ports.stream().map(this::createServicePort).toList())
+                .endSpec()
+                .build();
+    }
 
-        return builder;
+    void deleteDeployment(PluginCoordinates pluginCoordinates, Result result) {
+        List<StatusDetails> details = k8sClient.apps().deployments().withName(pluginCoordinates.getArtifactId()).delete();
+
+        if (details.isEmpty()) {
+            result.setSuccess(false);
+            result.getErrors().add("No deployment exists for " + pluginCoordinates.groupAndArtifact());
+        } else if(details.size() > 1) {
+            result.setSuccess(false);
+            result.getErrors().add("Unexpected delete results for deployment " + pluginCoordinates.getArtifactId() + " " + details);
+        }
+
+        checkStatusDetail(details.get(0), result);
+    }
+
+    void deleteService(PluginCoordinates pluginCoordinates, Result result) {
+        // try to delete a service, ignore the case where it does not exist (in most cases it won't)
+        List<StatusDetails> details = k8sClient.apps().deployments().withName(pluginCoordinates.getArtifactId()).delete();
+
+        if(details.size() > 1) {
+            result.setSuccess(false);
+            result.getErrors().add("Unexpected delete results for service " + pluginCoordinates.getArtifactId() + " " + details);
+        }
+
+        if (!details.isEmpty()) {
+            checkStatusDetail(details.get(0), result);
+        }
+    }
+
+    void checkStatusDetail(StatusDetails statusDetails, Result result) {
+        List<StatusCause> statusCauses = statusDetails.getCauses();
+
+        if (!statusCauses.isEmpty()) {
+            result.setSuccess(false);
+            result.getErrors().addAll(statusCauses.stream().map(this::statusCause).toList());
+        }
+    }
+
+    String statusCause(StatusCause statusCause) {
+        return "Field: " + statusCause.getField() + " Message: " + statusCause.getMessage() + " Reason: " + statusCause.getReason();
+    }
+
+    private ContainerPort createContainerPort(Integer port) {
+        return new ContainerPortBuilder()
+                .withContainerPort(port)
+                .withName("port-" + port)
+                .build();
+    }
+
+    private ServicePort createServicePort(Integer port) {
+        return new ServicePortBuilder()
+                .withPort(port)
+                .withTargetPort(new IntOrString(port))
+                .build();
     }
 
     /**
@@ -209,13 +296,85 @@ public class K8sDeployerService extends BaseDeployerService {
         deployment.getSpec().getTemplate().getSpec().getVolumes().add(configVolume);
     }
 
-    void preserveValuesIfUpgrade(Deployment upgradedDeployment, Deployment existingDeployment) {
+    void preserveValues(Deployment upgradedDeployment, Deployment existingDeployment) {
         if (existingDeployment != null) {
             upgradedDeployment.getSpec().setReplicas(existingDeployment.getSpec().getReplicas());
         }
     }
 
+    private boolean rolloutSuccessful(Deployment deployment) {
+        if (deployment == null) {
+            throw new IllegalStateException("Deployment was removed prior to completing");
+        }
+
+        DeploymentStatus deploymentStatus = deployment.getStatus();
+
+        if (deploymentStatus == null) {
+            log.debug("Waiting for deployment spec update to be observed...\n");
+            return false;
+        }
+
+        String deploymentName = deployment.getMetadata().getName();
+        long generation = getGeneration(deployment.getMetadata());
+        long observedGeneration = getObservedGeneration(deployment.getStatus());
+
+        Integer specReplicas = deployment.getSpec().getReplicas();
+        int statusReplicas = valueOrDefault(deploymentStatus.getReplicas(), 0);
+        int updatedReplicas = valueOrDefault(deploymentStatus.getUpdatedReplicas(), 0);
+        int availableReplicas = valueOrDefault(deploymentStatus.getAvailableReplicas(), 0);
+
+        if (generation <= observedGeneration) {
+            if (deployment.getStatus().getConditions().stream().anyMatch(this::progressingTimedOut)) {
+                throw new IllegalStateException("Deployment " + deploymentName + " exceeded its progress deadline");
+            }
+
+            if (specReplicas != null && updatedReplicas < specReplicas) {
+                log.debug("Waiting for deployment {} rollout to finish: {} out of {} new replicas have been updated...\n", deploymentName, updatedReplicas, specReplicas);
+                return false;
+            }
+
+            if (statusReplicas > updatedReplicas) {
+                log.debug("Waiting for deployment {} rollout to finish: {} old replicas are pending termination...\n", deploymentName, statusReplicas-updatedReplicas);
+                return false;
+            }
+
+            if (availableReplicas < updatedReplicas) {
+                log.debug("Waiting for deployment {} rollout to finish: {} of {} updated replicas are available...\n", deploymentName, availableReplicas, updatedReplicas);
+                return false;
+            }
+
+            return true;
+        }
+
+        return false;
+    }
+
+    private long getGeneration(ObjectMeta deploymentMetadata) {
+        return valueOrDefault(deploymentMetadata.getGeneration(), 0);
+    }
+
+    private long getObservedGeneration(DeploymentStatus status) {
+        return status != null ? valueOrDefault(status.getObservedGeneration(), -1) : -1;
+    }
+
+    private long valueOrDefault(Long value, long defaultValue) {
+        return value != null ? value : defaultValue;
+    }
+
+    private int valueOrDefault(Integer value, int defaultValue) {
+        return value != null ? value : defaultValue;
+    }
+
+    private boolean progressingTimedOut(DeploymentCondition condition) {
+        return PROGRESSING.equals(condition.getType()) && PROGRESS_DEADLINE_EXCEEDED.equals(condition.getReason());
+    }
+
     private Deployment loadBaseDeployment() throws IOException {
         return Serialization.unmarshal(baseDeployment.getInputStream(), Deployment.class);
+    }
+
+    private long getTimeoutInMillis() {
+        Duration timeout = deltaFiPropertiesService.getDeltaFiProperties().getPlugins().getDeployTimeout();
+        return timeout != null ? timeout.toMillis() : 60_000L;
     }
 }